Refactor: 안쓰는 코드 삭제
