+++ lib/common/provider/register_dialog_screen.dart
